Update tests with CInt

+++ tests/backends/ffi_import_01.c
@@ -20,10 +20,10 @@ void FFI_null_bits128(uint64_t x[2], uint64_t r[2]) { memcpy(r, x, 2*sizeof(uint
 const char* FFI_null_string(const char *x) { return x; }
 enum E FFI_null_E(enum E x) { return x; }
 bool FFI_null_Boolean(bool x) { return x; }
-int FFI_null_integer(int x) { return x; }
+int64_t FFI_null_integer(int64_t x) { return x; }
 int FFI_null_sint17(int x) { return x; }
 
-void FFI_int_bool(int x, int* ret1, bool* ret2)
+void FFI_int_bool(int64_t x, int64_t* ret1, bool* ret2)
 {
     *ret1 = x;
     *ret2 = x > 3;
