+++ tests/backends/ffi_import_02.c
@@ -6,28 +6,28 @@
 
 #include "isa_ffi.h"
 
-static int x = 0;
+static int64_t x = 0;
 
-int v1_Read(void)
+int64_t v1_Read(void)
 {
         return x + 1;
 }
 
-void v1_Write(int value)
+void v1_Write(int64_t value)
 {
         x = value + 1;
 }
 
 #define SIZE 1
-static int y[SIZE] = { 0 };
+static int64_t y[SIZE] = { 0 };
 
-int v2_Read(int i)
+int64_t v2_Read(int64_t i)
 {
         assert(i >= 0 && i < SIZE);
         return y[i] + 1;
 }
 
-void v2_Write(int i, int value)
+void v2_Write(int64_t i, int64_t value)
 {
         assert(i >= 0 && i < SIZE);
         y[i] = value + 1;
